skylab: a switch to only talk to default inventory
This switch actually only affect `remove-dut` and `batch-update-dut`
which bypass crosskylabadmin to operate inventory. When enable this
switch, skylab will only write to default inventory.

BUG=962689
TEST=Ran skylab command against cros-lab-inventory-dev

Reviewed-on: https://chromium-review.googlesource.com/c/infra/infra/+/2077798
Auto-Submit: Congbin Guo <guocb@chromium.org>
Commit-Queue: Congbin Guo <guocb@chromium.org>
Cr-Commit-Position: refs/heads/master@{#29349}

+++ go/src/infra/cmd/skylab/internal/cmd/inventory/remove-duts.go
@@ -98,13 +98,19 @@ func (c *removeDutsRun) innerRun(a subcommands.Application, args []string, env s
 	icMain := NewInventoryClient(hc, e, true)
 	icBackup := NewInventoryClient(hc, e, false)
 
-	icBackup.removeDUTs(ctx, c.server, c.Flags.Args(), c.removalReason, a.GetOut())
+	if e.DefaultInventoryOnly {
+		fmt.Fprintln(a.GetOut(), "= Skip the operation on backup inventory system. =")
+	} else {
+		icBackup.removeDUTs(ctx, c.server, c.Flags.Args(), c.removalReason, a.GetOut())
+	}
 	modified, err := icMain.removeDUTs(ctx, c.server, c.Flags.Args(), c.removalReason, a.GetOut())
 	if err != nil {
 		return err
 	}
 	if c.delete {
-		icBackup.deleteDUTs(ctx, c.Flags.Args(), &c.authFlags, a.GetOut())
+		if !e.DefaultInventoryOnly {
+			icBackup.deleteDUTs(ctx, c.Flags.Args(), &c.authFlags, a.GetOut())
+		}
 		mod, err := icMain.deleteDUTs(ctx, c.Flags.Args(), &c.authFlags, a.GetOut())
 		if err != nil {
 			return err
